Split of QoL features into separate files
