+++ Source/qol/autopickup.h
@@ -1,0 +1,13 @@
+/**
+* @file autopickup.h
+*
+* QoL feature for automatically picking up gold
+*/
+
+#pragma once
+
+namespace devilution {
+
+void AutoGoldPickup(int pnum);
+
+} // namespace devilution
